OOP版LSコマンドほぼ完成。クラスごとにファイルを分けた。
